merge components CustomBackButton and CustomBackButtonWithSideElement into one component CustomBackButton
The component now can be used without title or optionally with title/customElement on the right edge next to the back button.
All Screens using either one of the two components were refactored to use the CustomBackButton in the manner needed by the specific screen. The paddings of the outer container were adjusted where needed.